+++ test/110_hap_identity.js
@@ -23,7 +23,7 @@ describe('HomeKit-CCU hapIdentity', () => {
   })
 
   it('accepts only well formed, non trivial setup codes', () => {
-    expect(isValidSetupCode('031-45-154')).to.be(true)
+    expect(isValidSetupCode('482-91-736')).to.be(true)
     expect(isValidSetupCode('123-45-678')).to.be(false)
     expect(isValidSetupCode('876-54-321')).to.be(false)
     expect(isValidSetupCode('000-00-000')).to.be(false)